# Core: Optimize computing user-facing state in data task