--- a/api/src/main/java/org/apache/iceberg/BaseScanTaskGroup.java
+++ b/api/src/main/java/org/apache/iceberg/BaseScanTaskGroup.java
@@ -19,7 +19,7 @@
 package org.apache.iceberg;
 
 import java.util.Collection;
-import java.util.List;
+import java.util.Collections;
 import org.apache.iceberg.relocated.com.google.common.base.Joiner;
 import org.apache.iceberg.relocated.com.google.common.base.MoreObjects;
 import org.apache.iceberg.relocated.com.google.common.base.Preconditions;
@@ -28,12 +28,13 @@ import org.apache.iceberg.relocated.com.google.common.collect.ImmutableList;
 public class BaseScanTaskGroup<T extends ScanTask> implements ScanTaskGroup<T> {
   private final StructLike groupingKey;
   private final Object[] tasks;
-  private transient volatile List<T> taskList;
+  private transient volatile Collection<T> taskCollection;
 
   public BaseScanTaskGroup(StructLike groupingKey, Collection<T> tasks) {
     Preconditions.checkNotNull(tasks, "tasks cannot be null");
     this.groupingKey = groupingKey;
     this.tasks = tasks.toArray();
+    this.taskCollection = Collections.unmodifiableCollection(tasks);
   }
 
   public BaseScanTaskGroup(Collection<T> tasks) {
@@ -48,20 +49,47 @@ public class BaseScanTaskGroup<T extends ScanTask> implements ScanTaskGroup<T> {
   @Override
   @SuppressWarnings("unchecked")
   public Collection<T> tasks() {
-    if (taskList == null) {
+    if (taskCollection == null) {
       synchronized (this) {
-        if (taskList == null) {
+        if (taskCollection == null) {
           ImmutableList.Builder<T> listBuilder =
               ImmutableList.builderWithExpectedSize(tasks.length);
           for (Object task : tasks) {
             listBuilder.add((T) task);
           }
-          taskList = listBuilder.build();
+          this.taskCollection = listBuilder.build();
         }
       }
     }
 
-    return taskList;
+    return taskCollection;
+  }
+
+  @Override
+  public long sizeBytes() {
+    long sizeBytes = 0L;
+    for (Object task : tasks) {
+      sizeBytes += ((ScanTask) task).sizeBytes();
+    }
+    return sizeBytes;
+  }
+
+  @Override
+  public long estimatedRowsCount() {
+    long estimatedRowsCount = 0L;
+    for (Object task : tasks) {
+      estimatedRowsCount += ((ScanTask) task).estimatedRowsCount();
+    }
+    return estimatedRowsCount;
+  }
+
+  @Override
+  public int filesCount() {
+    int filesCount = 0;
+    for (Object task : tasks) {
+      filesCount += ((ScanTask) task).filesCount();
+    }
+    return filesCount;
   }
 
   @Override