--- a/core/src/main/java/org/apache/iceberg/BaseFileScanTask.java
+++ b/core/src/main/java/org/apache/iceberg/BaseFileScanTask.java
@@ -27,6 +27,8 @@ import org.apache.iceberg.relocated.com.google.common.collect.ImmutableList;
 public class BaseFileScanTask extends BaseContentScanTask<FileScanTask, DataFile>
     implements FileScanTask {
   private final DeleteFile[] deletes;
+  private transient volatile List<DeleteFile> deleteList = null;
+  private transient volatile long deletesSizeBytes = 0L;
 
   public BaseFileScanTask(
       DataFile file,
@@ -45,12 +47,26 @@ public class BaseFileScanTask extends BaseContentScanTask<FileScanTask, DataFile
 
   @Override
   protected FileScanTask newSplitTask(FileScanTask parentTask, long offset, long length) {
-    return new SplitScanTask(offset, length, parentTask);
+    return new SplitScanTask(offset, length, parentTask, deletesSizeBytes());
   }
 
   @Override
   public List<DeleteFile> deletes() {
-    return ImmutableList.copyOf(deletes);
+    if (deleteList == null) {
+      this.deleteList = ImmutableList.copyOf(deletes);
+    }
+
+    return deleteList;
+  }
+
+  @Override
+  public long sizeBytes() {
+    return length() + deletesSizeBytes();
+  }
+
+  @Override
+  public int filesCount() {
+    return 1 + deletes.length;
   }
 
   @Override
@@ -58,16 +74,37 @@ public class BaseFileScanTask extends BaseContentScanTask<FileScanTask, DataFile
     return super.schema();
   }
 
+  // lazily cache the size of deletes to reuse in all split tasks
+  private long deletesSizeBytes() {
+    if (deletesSizeBytes == 0L && deletes.length > 0) {
+      long size = 0L;
+      for (DeleteFile deleteFile : deletes) {
+        size += deleteFile.fileSizeInBytes();
+      }
+      this.deletesSizeBytes = size;
+    }
+
+    return deletesSizeBytes;
+  }
+
   @VisibleForTesting
   static final class SplitScanTask implements FileScanTask, MergeableScanTask<SplitScanTask> {
     private final long len;
     private final long offset;
     private final FileScanTask fileScanTask;
+    private transient volatile long deletesSizeBytes = 0L;
 
     SplitScanTask(long offset, long len, FileScanTask fileScanTask) {
       this.offset = offset;
       this.len = len;
       this.fileScanTask = fileScanTask;
+    }
+
+    SplitScanTask(long offset, long len, FileScanTask fileScanTask, long deletesSizeBytes) {
+      this.offset = offset;
+      this.len = len;
+      this.fileScanTask = fileScanTask;
+      this.deletesSizeBytes = deletesSizeBytes;
     }
 
     @Override
@@ -106,6 +143,16 @@ public class BaseFileScanTask extends BaseContentScanTask<FileScanTask, DataFile
     }
 
     @Override
+    public long sizeBytes() {
+      return len + deletesSizeBytes();
+    }
+
+    @Override
+    public int filesCount() {
+      return fileScanTask.filesCount();
+    }
+
+    @Override
     public Expression residual() {
       return fileScanTask.residual();
     }
@@ -128,7 +175,19 @@ public class BaseFileScanTask extends BaseContentScanTask<FileScanTask, DataFile
     @Override
     public SplitScanTask merge(ScanTask other) {
       SplitScanTask that = (SplitScanTask) other;
-      return new SplitScanTask(offset, len + that.length(), fileScanTask);
+      return new SplitScanTask(offset, len + that.length(), fileScanTask, deletesSizeBytes);
+    }
+
+    private long deletesSizeBytes() {
+      if (deletesSizeBytes == 0L && fileScanTask.filesCount() > 1) {
+        long size = 0L;
+        for (DeleteFile deleteFile : fileScanTask.deletes()) {
+          size += deleteFile.fileSizeInBytes();
+        }
+        this.deletesSizeBytes = size;
+      }
+
+      return deletesSizeBytes;
     }
   }
 }