This PR optimizes computing user-facing state in data tasks to reduce the garbage and improve performance. 

It is related to the work in #8336.