@aokolnychyi 2023-08-23
The behavior should be exactly like before. There were some some tests for planning as well as for Spark and Flink serialization. I'll find and post them tomorrow.