@aokolnychyi 2023-08-24
There are tests in `TestDataTableScan`, `TestBatchScans` and in `TestTableScanUtil` + each engine has serialization tests. I feel like they cover most of what we need.